Solr port fix
The ports docker option only applies to external networking.
The solr app in the container still runs on it's default port.
External access has port rewriting by Docker.
But internal container to container networking has to go to the port the app is listening on
(and which is EXPOSEd by the container by default, so no explicit EXPOSE needed).
So need to hard code default port in config var or "python src/handler.py -t solrize" can't reach Solr and it fails

The same is not true for Postgres as setting PGPORT actually changes the port the server runs on.

Also copyied and pasted comment about Solr port can be changed